==> e21_momentum_investing_using_yfinance/0_get_data.ipynb <==
import yfinance as yf
import os

stock_names = []
for dirpath, subdir, files in os.walk('../e20_sensex_generating buy signals using linear regression/data/individual stocks/'): 
    for i in files:
        stock_names.append(i.split('-EQ-')[0].split('-')[-1] + '.NS')

stock_names.append("^BSESN")

for i in stock_names:
    df = yf.download(i, start='2021-06-01', end='2023-06-02', group_by='ticker')
    df.reset_index(inplace = True)
    df.to_csv('stock_prices/' + i.replace('^', '') + '.csv', index = False)
