==> e21_momentum_investing_using_yfinance/2_individual stocks.ipynb <==
import pandas as pd
import dateutil
from datetime import datetime
import statistics as st
import os

df_list = []
stock_names = []

for dirpath, subdir, files in os.walk('./data/individual stocks/'): 
    for i in files:
        df = pd.read_csv(dirpath + i, index_col= False)
        df_list.append(df)
        stock_names.append(i)

stock_names

df_list[0].head()

def get_monthly_slopes(df, stock_name):
    df['month'] = df['Date'].apply(lambda x: str(dateutil.parser.parse(x).year) + "-" 
                                    + str(dateutil.parser.parse(x).month).zfill(2))
    month = df['month'].unique()
    try:
        df['close '] = df['Close'].apply(lambda x: float(x.replace(',', '')))
    except AttributeError as e:
        pass
    
    months = []
    slopes = []

    for i in month:
        x = list(df[df['month'] == i].index)
        y = list(df[df['month'] == i]['Close'])

        try:
            slope, intercept = st.linear_regression(x, y)
            slopes.append(slope)
            months.append(i)
        except st.StatisticsError as e:
            print(e)

    df_out = pd.DataFrame({
        'month': months,
        stock_name: slopes
    })
    return df_out

df_monthly_slopes = [get_monthly_slopes(df = i[0], stock_name = i[1]) for i in zip(df_list, stock_names)]

df_rep = df_monthly_slopes[0]
for i in df_monthly_slopes[1:]:
    df_rep = pd.merge(df_rep, i, on = 'month', how = 'inner')

df_rep.sort_values(by='month')

df_rep.to_csv('out/monthly_slopes.csv', index = False)

df_rep.style.applymap(lambda x: "background-color: red" if x>0 else "background-color: white")
