==> e21_momentum_investing_using_yfinance/4_generate txns for sensex.ipynb <==
import pandas as pd
import dateutil
from datetime import datetime
import statistics as st
from pandas.tseries.offsets import Day, MonthEnd

df = pd.read_csv('data/BSESN.csv', index_col=False)

df['Date'] = df['Date'].apply(lambda x: dateutil.parser.parse(x).strftime("%d-%b-%Y"))

df

start_date = dateutil.parser.parse('1-July-2021')
end_date = dateutil.parser.parse('1-June-2023')

# MS: month begin
# BMS: BusinessMonthBegin
dr = pd.date_range(start_date, end_date, freq="BMS")

stock = 'data/BSESN.csv'
amount = 1000

stock_l = []
amt_l = []
close_l = []
date_l = []

for i in dr:
    j = i
    close = df[df['Date'] == j.strftime("%d-%b-%Y")]['Close']
    while len(close) == 0:
        j = i + Day()
        close = df[df['Date'] == j.strftime("%d-%b-%Y")]['Close']
        
    
    stock_l.append(stock)
    amt_l.append(amount)
    close_l.append(list(close.values)[0])
    date_l.append(i)

df_rep = pd.DataFrame({
    'stock': stock_l,
    'amt': amt_l,
    'close': close_l,
    'date': date_l
})

df_rep

df_rep.to_csv('out/sensex_txns.csv', index = False)
