==> e21_momentum_investing_using_yfinance/6_gain_from_sensex.ipynb <==
import pandas as pd

# Calculate Gain From SIP in Sensex

df_sensex = pd.read_csv('data/BSESN.csv', index_col = False)

df_sensex

current_close = df_sensex[df_sensex['Date'] == '2023-06-01']['Close'].values[0]

current_close

df_txns_sensex = pd.read_csv('out/sensex_txns.csv')

df_txns_sensex = df_txns_sensex[df_txns_sensex['date'] != '2023-06-01']

def get_current_value(row):
    curr_value = row['amt'] * (current_close / row['close'])
    return curr_value

df_txns_sensex['curr_value'] = df_txns_sensex.apply(get_current_value, axis = 1)

df_txns_sensex

total_invested = sum(df_txns_sensex['amt'].values)
current_value = sum(df_txns_sensex['curr_value'].values)

print("Amount invested:", total_invested)
print("current_value:", round(current_value, 2))
print("Percentage gain:", round((current_value - total_invested)/total_invested * 100, 2))
