==> e21_momentum_investing_using_yfinance/7_gain_from_stocks.ipynb <==
import pandas as pd

df_t = pd.read_csv('data/individual stocks/ASIANPAINT.NS.csv')

df_t

# Calculate Gain From Momentum Investing Based SIP in Sensex

df_txns = pd.read_csv('out/stock_txns.csv')
df_txns = df_txns[df_txns['txn_date'] != '2023-06-01']

df_txns

current_close_l = []
current_value_l = []

date_of_evaluation = '2023-06-01'

def get_current_value(row):
    stock_filename = "data/individual stocks/" + row['stock']
    
    df = pd.read_csv(stock_filename)
    
    current_close = df[df['Date'] == date_of_evaluation]['Close']
    
    try:
        current_close = current_close.values[0].replace(",", "")
    except:
        current_close = current_close.values[0]
    
    try: 
        txn_close = float(row['close'].replace(',', ''))
    except:
        txn_close = float(row['close'])
        
    current_value = row['amt'] * (float(current_close) / txn_close)
    
    current_close_l.append(current_close)
    current_value_l.append(current_value)
    
_ = df_txns.apply(get_current_value, axis = 1)

df_txns['curr_value'] = current_value_l
df_txns['curr_close'] = current_close_l

df_txns

total_invested = sum(df_txns['amt'].values)
current_value = sum(df_txns['curr_value'].values)

print("Amount invested:", total_invested)
print("current_value:", round(current_value, 2))
print("Percentage gain:", round((current_value - total_invested)/total_invested * 100, 2))
