==> e21_momentum_investing_using_yfinance/8_generate sell signals.ipynb <==
import pandas as pd
import dateutil
from pandas.tseries.offsets import Day, BusinessMonthBegin, MonthEnd

monthly_slopes = pd.read_csv('out/monthly_slopes.csv')

monthly_slopes

def generate_sell_signals(monthly_slopes):
    months = monthly_slopes['month'].values
    cnt = 0
    
    df_month = []
    df_stock = []
    
    for m in months:
        slopes = monthly_slopes[monthly_slopes['month'] == m].to_dict()
        slopes.pop('month')
        
        stocks_sorted_by_slopes = dict(sorted(slopes.items(), key=lambda x:x[1][cnt], reverse = True))
        stocks_sorted_by_slopes = {k:v for k,v in stocks_sorted_by_slopes.items() if v[cnt] < 0}
        cnt += 1
        
        _ = [df_month.append(m) for i in stocks_sorted_by_slopes]
        _ = [df_stock.append(i) for i in stocks_sorted_by_slopes]
        
        
    df_rep = pd.DataFrame({
        "past_month": df_month,
        "stock": df_stock
    })
    return df_rep

df = generate_sell_signals(monthly_slopes)

df

txn_dates = []
close_l = []

def get_txn_date_and_close(row):    
    txn_date = row['past_month']
    txn_date += '-01'
    txn_date = dateutil.parser.parse(txn_date) + MonthEnd() + Day()
    
    df = pd.read_csv('data/individual stocks/' + row['stock'])
    df['Date '] = df['Date'].apply(lambda x: dateutil.parser.parse(x).strftime("%Y-%m-%d"))
    
    txn_dates.append(txn_date)
    
    j = txn_date + Day()
    close = df[df['Date'] == str(txn_date).split()[0]]['Close'].values
    while len(close) == 0:
        j += Day()
        close = df[df['Date'] == str(j).split()[0]]['Close'].values
        
    close_l.append(close[0])
    
_ = df.apply(get_txn_date_and_close, axis = 1)

df['txn_date'] = txn_dates
df['close'] = close_l

df

df.to_csv('out/sell_signals.csv', index = False)
