==> e21_momentum_investing_using_yfinance/9_gen_txns_with_buy_and_sell.ipynb <==
import pandas as pd
from datetime import datetime 

df_stock_txns_with_buys = pd.read_csv('out/stock_txns.csv')

df_stock_txns_with_buys.head()

df_sell_signals = pd.read_csv('out/sell_signals.csv')

df_sell_signals.head()

df_stock_txns_with_buys.shape

sell_dates = []
sell_nav = []

def populate_sell_txns(row):
    df = df_sell_signals[
        (df_sell_signals['stock'] == row['stock']) 
        & (df_sell_signals['txn_date'] > row['txn_date'])
    ]
    
    df.sort_values(by='txn_date', inplace = True)
    
    if(len(df) > 0):
        sell_dates.append(df['txn_date'].values[0])
        sell_nav.append(df['close'].values[0])
    else:
        sell_dates.append('Not Applicable')
        sell_nav.append('Not Applicable')
        
_ = df_stock_txns_with_buys.apply(populate_sell_txns, axis = 1)
df_stock_txns_with_buys['sell_dates'] = sell_dates
df_stock_txns_with_buys['sell_nav'] = sell_nav

df_stock_txns_with_buys.to_csv('out/stock_txns_with_buys_and_sells.csv', index = False)
