==> e21_momentum_investing_using_yfinance/11_gain_from_stocks after reinvesting.ipynb <==
import pandas as pd

df_txns = pd.read_csv('out/buy_sell.csv')

df_txns.drop(columns=['sell_amt'], inplace = True)

df_txns

buy_dates = sorted(list(df_txns['txn_date'].unique()))

next_buy_amts = {}

def increment_buy_amt(row):
    sell_date = row['sell_dates']
    
    try:
        next_date = buy_dates[buy_dates.index(sell_date) + 1]
    
        if(next_date in next_buy_amts):
            next_buy_amts[next_date] += 40
        else:
            next_buy_amts[next_date] = df_txns[df_txns['txn_date'] == next_date]['amt'].values[0] + 40
            
        return True
    except:
        return False

_ = df_txns.apply(increment_buy_amt, axis = 1)

next_buy_amts

def update_buy_amt(row):
    if(row['txn_date'] in next_buy_amts):
        amt = next_buy_amts[row['txn_date']]
    else:
        amt = row['amt']
    return amt

df_txns['amt'] = df_txns.apply(update_buy_amt, axis = 1)

df_txns.head(20)

date_of_evaluation = '2023-06-01'
df_txns = df_txns[df_txns['txn_date'] != date_of_evaluation]
def get_sell_nav(row):
    if(row['sell_nav'] == 'Not Applicable'):
        stock_filename = "data/individual stocks/" + row['stock']

        df = pd.read_csv(stock_filename)

        sell_nav = df[df['Date'] == date_of_evaluation]['Close']

        return float(sell_nav.values[0])
    else:
        return float(row['sell_nav'])
        
        
df_txns['sell_nav'] = df_txns.apply(get_sell_nav, axis = 1)

df_txns['sell_dates'] = df_txns['sell_dates'].apply(lambda x: x if x != 'Not Applicable' else date_of_evaluation)

def gen_sell_amt(row):
    return row['amt'] * row['sell_nav'] / row['close']

df_txns['sell_amt'] = df_txns.apply(gen_sell_amt, axis = 1)

df_txns

total_invested = sum(df_txns['amt'].values)
current_value = sum(df_txns['sell_amt'].values)

print("Amount invested:", total_invested)
print("current_value:", round(current_value, 2))
print("Percentage gain:", round((current_value - total_invested)/total_invested * 100, 2))
